Change name to Divisor
